cl-adm/cmd: rename cl-adm CLI to clusterlink
